refactor(state): move LoreStore interface to loreStore.types.ts
loreStore.actions.ts typed its factories against the LoreStore interface
imported from loreStore.ts, which itself imports the factories — a type-only
cycle. The interface now lives in loreStore.types.ts and both sides import
from there.
